Allow passing in symbols as non-autodiff parameters.
They will not be autodiffed, but it doesn't mean we can't recycle
there values. This is useful for better reuse of vjp and jvp models.

After this change I think we can completely remove all ains of vjp
and jvp operators, since we do not intend to take derivatives of
them anyways.