Performance improvements by reorg of imports.
It turns out that the workaround for circular imports is taking
quite a bit of time as it is called in hot functions.

The dependency on stdlib and autodiff are now routed through
vmad.core.get_stdlib() and vmad.core.get_autodiff()

I also reordered the import lines at header of each file.